Use wgpu-matrix 2.x
The braking change was in 1.x, mat4.lookAt returned a camera matrix
which was unexpected by most devs. 2.x changed mat4.lookAt to return
a view matrix so that devs would get what they expect coming from
other math libraries.

+++ src/sample/samplerParameters/main.ts
@@ -280,7 +280,7 @@ const init: SampleInit = async ({ canvas, pageState, gui }) => {
   const viewProj = mat4.translate(
     mat4.perspective(2 * Math.atan(1 / kCameraDist), 1, 0.1, 100),
     [0, 0, -kCameraDist]
-  );
+  ) as Float32Array;
   device.queue.writeBuffer(bufConfig, 0, viewProj);
 
   const bufMatrices = device.createBuffer({
